assets/requirements.txt
backcall==0.2.0
decorator==4.4.2
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
numpy==1.19.4
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
prompt-toolkit==3.0.8
ptyprocess==0.6.0
Pygments==2.7.3
torch==1.7.1+cpu
torchaudio==0.7.2
torchvision==0.8.2+cpu
traitlets==5.0.5
typing-extensions==3.7.4.3
wcwidth==0.2.5